With this checkin I am preparing to debug indexing on numeric lists and to add functionality to resolve ranges in an index.
I also removed a lot of redundant 'DEB' comments. For example, 
addIDs(CIfol, IDfol) doesn't need "add ids IDfol to CI's follower"
Some of the comments were out of date anyhow.

These comments were intended to be used for debugging but it's too complex
and not useful. So far, using ddd to debug is the only practical solution.